Correct crossfeed filter response
Since we're filtering the original sound to compensate for the frequency
rolloff in the crossfeed filter, the crossfeed response is effectively the
square of the transfer function, so let's take its square root so the
response we get is what we actually wanted.

This also modifies the filter designer to minimize the square of the
filter's error, which should cause it to limit max deviation from the ideal
response.